First-person 2: Movement
had to revamp the movement handling of the game (both top-down and first-person) to use character controller component. This ensures that player movement is smooth (it is jittery if I try to alternate between both character controller and rigidbody). It also ensures that gravity works on the player in both modes of view.

As a sacrafice, movement for top-down is near identical to the old movement.